+++ CHANGELOG.md
@@ -7,6 +7,7 @@ It is a summary of changes that would be pertinent to the end user of the PGEtin
 ## 2024-06-09
 
 - Fixed [Issue #92](https://github.com/Moros1138/PGEtinker/issues/92)
+- Fixed default layout button handler
 
 ## 2024-06-08